cart Service completed

+++ src/main/java/com/bikkadit/electronicstore/dtos/CartItemDto.java
@@ -1,13 +1,6 @@
 package com.bikkadit.electronicstore.dtos;
-
-import com.bikkadit.electronicstore.entities.Cart;
 import com.bikkadit.electronicstore.entities.Product;
 import lombok.*;
-
-import javax.persistence.FetchType;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.OneToOne;
 @Setter
 @Getter
 @NoArgsConstructor
@@ -21,5 +14,5 @@ public class CartItemDto {
 
     private int quantity;
 
-    private int totalPrice;
+    private double totalPrice;
 }
